+++ sdk/textanalytics/azure-ai-textanalytics/azure/ai/textanalytics/aio/_text_analytics_client_async.py
@@ -50,7 +50,9 @@ from .._models import (
     SingleLabelClassifyAction,
     MultiLabelClassifyAction,
     ClassifyDocumentResult,
-    AnalyzeHealthcareEntitiesAction
+    AnalyzeHealthcareEntitiesAction,
+    ExtractSummaryAction,
+    ExtractSummaryResult,
 )
 from .._check import is_language_api, string_index_type_compatibility
 from .._lro import TextAnalyticsOperationResourcePolling
@@ -75,6 +77,7 @@ AsyncAnalyzeActionsResponse = AsyncTextAnalysisLROPoller[
                 RecognizeCustomEntitiesResult,
                 ClassifyDocumentResult,
                 AnalyzeHealthcareEntitiesResult,
+                ExtractSummaryResult,
                 DocumentError,
             ]
         ]
@@ -831,7 +834,10 @@ class TextAnalyticsClient(AsyncTextAnalyticsClientBase):
     @distributed_trace_async
     @validate_multiapi_args(
         version_method_added="v3.1",
-        args_mapping={"2022-05-01": ["display_name"]}
+        args_mapping={
+            "2022-10-01-preview": ["fhir_version", "document_type"],
+            "2022-05-01": ["display_name"]
+        }
     )
     async def begin_analyze_healthcare_entities(
         self,
@@ -882,6 +888,14 @@ class TextAnalyticsClient(AsyncTextAnalyticsClientBase):
             Cognitive Services Compliance and Privacy notes at https://aka.ms/cs-compliance for
             additional details, and Microsoft Responsible AI principles at
             https://www.microsoft.com/ai/responsible-ai.
+        :keyword str fhir_version: The FHIR Spec version that the result will use to format the fhir_bundle
+            on the result object. For additional information see https://www.hl7.org/fhir/overview.html.
+            The only acceptable values to pass in are None and "4.0.1". The default value is None.
+        :keyword document_type: Document type that can be provided as input for Fhir Documents. Expect to
+            have fhir_version provided when used. Behavior of using None enum is the same as not using the
+            document_type parameter. Known values are: "None", "ClinicalTrial", "DischargeSummary",
+            "ProgressNote", "HistoryAndPhysical", "Consult", "Imaging", "Pathology", and "ProcedureNote".
+        :paramtype document_type: str or ~azure.ai.textanalytics.HealthcareDocumentType
         :return: An instance of an AsyncAnalyzeHealthcareEntitiesLROPoller. Call `result()` on the poller
             object to return a heterogeneous pageable of
             :class:`~azure.ai.textanalytics.AnalyzeHealthcareEntitiesResult` and
@@ -895,6 +909,8 @@ class TextAnalyticsClient(AsyncTextAnalyticsClientBase):
             The *begin_analyze_healthcare_entities* client method.
         .. versionadded:: 2022-05-01
             The *display_name* keyword argument.
+        .. versionadded:: 2022-10-01-preview
+            The *fhir_version* and *document_type* keyword arguments.
 
         .. admonition:: Example:
 
@@ -914,6 +930,8 @@ class TextAnalyticsClient(AsyncTextAnalyticsClientBase):
         string_index_type = kwargs.pop("string_index_type", self._string_code_unit)
         disable_service_logs = kwargs.pop("disable_service_logs", None)
         display_name = kwargs.pop("display_name", None)
+        fhir_version = kwargs.pop("fhir_version", None)
+        document_type = kwargs.pop("document_type", None)
 
         if continuation_token:
             return cast(
@@ -963,6 +981,8 @@ class TextAnalyticsClient(AsyncTextAnalyticsClientBase):
                                         model_version=model_version,
                                         logging_opt_out=disable_service_logs,
                                         string_index_type=string_index_type_compatibility(string_index_type),
+                                        fhir_version=fhir_version,
+                                        document_type=document_type,
                                     )
                                 )
                             ]
@@ -1053,6 +1073,7 @@ class TextAnalyticsClient(AsyncTextAnalyticsClientBase):
                 SingleLabelClassifyAction,
                 MultiLabelClassifyAction,
                 AnalyzeHealthcareEntitiesAction,
+                ExtractSummaryAction,
             ]
         ],
         **kwargs: Any,
@@ -1068,6 +1089,7 @@ class TextAnalyticsClient(AsyncTextAnalyticsClientBase):
                     RecognizeCustomEntitiesResult,
                     ClassifyDocumentResult,
                     AnalyzeHealthcareEntitiesResult,
+                    ExtractSummaryResult,
                     DocumentError,
                 ]
             ]
@@ -1096,7 +1118,7 @@ class TextAnalyticsClient(AsyncTextAnalyticsClientBase):
             list[RecognizeEntitiesAction or RecognizePiiEntitiesAction or ExtractKeyPhrasesAction or
             RecognizeLinkedEntitiesAction or AnalyzeSentimentAction or
             RecognizeCustomEntitiesAction or SingleLabelClassifyAction or
-            MultiLabelClassifyAction or AnalyzeHealthcareEntitiesAction]
+            MultiLabelClassifyAction or AnalyzeHealthcareEntitiesAction or ExtractSummaryAction]
         :keyword str display_name: An optional display name to set for the requested analysis.
         :keyword str language: The 2 letter ISO 639-1 representation of language for the
             entire batch. For example, use "en" for English; "es" for Spanish etc.
@@ -1125,7 +1147,7 @@ class TextAnalyticsClient(AsyncTextAnalyticsClientBase):
             ~azure.ai.textanalytics.aio.AsyncTextAnalysisLROPoller[~azure.core.async_paging.AsyncItemPaged[
             list[RecognizeEntitiesResult or RecognizeLinkedEntitiesResult or RecognizePiiEntitiesResult or
             ExtractKeyPhrasesResult or AnalyzeSentimentResult or RecognizeCustomEntitiesResult
-            or ClassifyDocumentResult or AnalyzeHealthcareEntitiesResult or DocumentError]]]
+            or ClassifyDocumentResult or AnalyzeHealthcareEntitiesResult or ExtractSummaryResult or DocumentError]]]
         :raises ~azure.core.exceptions.HttpResponseError or TypeError or ValueError:
 
         .. versionadded:: v3.1
@@ -1135,6 +1157,9 @@ class TextAnalyticsClient(AsyncTextAnalyticsClientBase):
             *MultiLabelClassifyAction*, and *AnalyzeHealthcareEntitiesAction* input options and the
             corresponding *RecognizeCustomEntitiesResult*, *ClassifyDocumentResult*,
             and *AnalyzeHealthcareEntitiesResult* result objects
+        .. versionadded:: 2022-10-01-preview
+            The *ExtractSummaryAction* input option and the corresponding *ExtractSummaryResult*
+            result object.
 
         .. admonition:: Example: